+++ src/itest/java/ru/hh/consul/util/CustomBuilderITest.java
@@ -1,0 +1,26 @@
+package ru.hh.consul.util;
+
+import ru.hh.consul.BaseIntegrationTest;
+import ru.hh.consul.Consul;
+import ru.hh.consul.model.agent.Agent;
+import java.net.Proxy;
+import java.net.UnknownHostException;
+import static org.junit.Assert.assertNotNull;
+import org.junit.Test;
+
+public class CustomBuilderITest extends BaseIntegrationTest{
+
+    @Test
+    public void shouldConnectWithCustomTimeouts() throws UnknownHostException {
+        Consul client = Consul.builder()
+                .withHostAndPort(defaultClientHostAndPort)
+                .withProxy(Proxy.NO_PROXY)
+                .withConnectTimeoutMillis(10000)
+                .withReadTimeoutMillis(3600000)
+                .withWriteTimeoutMillis(900)
+                .build();
+        Agent agent = client.agentClient().getAgent();
+        assertNotNull(agent);
+    }
+
+}