+++ src/main/java/ru/hh/consul/model/EventResponse.java
@@ -1,0 +1,20 @@
+package ru.hh.consul.model;
+
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import ru.hh.consul.model.event.Event;
+import java.math.BigInteger;
+import java.util.List;
+import org.immutables.value.Value;
+
+@Value.Immutable
+@JsonSerialize(as = ImmutableEventResponse.class)
+@JsonDeserialize(as = ImmutableEventResponse.class)
+public abstract class EventResponse {
+
+    @Value.Parameter
+    public abstract List<Event> getEvents();
+    @Value.Parameter
+    public abstract BigInteger getIndex();
+
+}