+++ src/main/java/ru/hh/consul/model/query/ServiceQuery.java
@@ -1,0 +1,27 @@
+package ru.hh.consul.model.query;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import java.util.List;
+import java.util.Optional;
+
+@org.immutables.value.Value.Immutable
+@JsonDeserialize(as = ImmutableServiceQuery.class)
+@JsonSerialize(as = ImmutableServiceQuery.class)
+@JsonIgnoreProperties(ignoreUnknown = true)
+public abstract class ServiceQuery {
+
+    @JsonProperty("Service")
+    public abstract String getService();
+
+    @JsonProperty("OnlyPassing")
+    public abstract Optional<Boolean> getOnlyPassing();
+
+    @JsonProperty("Tags")
+    public abstract Optional<List<String>> getTags();
+
+    @JsonProperty("Failover")
+    public abstract Optional<Failover> getFailover();
+}