+++ src/main/java/ru/hh/consul/option/DeleteOptions.java
@@ -1,0 +1,39 @@
+package ru.hh.consul.option;
+
+import static ru.hh.consul.option.Options.optionallyAdd;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+import org.immutables.value.Value;
+
+@Value.Immutable
+public abstract class DeleteOptions implements ParamAdder {
+
+	public static final DeleteOptions BLANK = ImmutableDeleteOptions.builder().build();
+	public static final DeleteOptions RECURSE = ImmutableDeleteOptions.builder().recurse(true).build();
+
+	public abstract Optional<Long> getCas();
+
+	public abstract Optional<Boolean> getRecurse();
+
+	public abstract Optional<String> getDatacenter();
+
+	@Value.Derived
+	public boolean isRecurse() {
+		return getRecurse().isPresent();
+	}
+
+	@Override
+	public Map<String, Object> toQuery() {
+		final Map<String, Object> result = new HashMap<>();
+
+		if (isRecurse()) {
+			result.put("recurse", "");
+		}
+
+		optionallyAdd(result, "cas", getCas());
+		optionallyAdd(result, "dc", getDatacenter());
+
+		return result;
+	}
+}