+++ src/main/java/ru/hh/consul/option/ParamAdder.java
@@ -1,0 +1,13 @@
+package ru.hh.consul.option;
+
+import java.util.List;
+import java.util.Map;
+
+public interface ParamAdder {
+
+    Map<String, Object> toQuery();
+
+    List<String> toQueryParameters();
+
+    Map<String, String> toHeaders();
+}