+++ src/main/java/ru/hh/consul/util/bookend/ConsulBookendContext.java
@@ -1,0 +1,27 @@
+package ru.hh.consul.util.bookend;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+
+public class ConsulBookendContext {
+
+    private Map<String, Object> data;
+
+    ConsulBookendContext() {
+
+    }
+
+    public void put(String key, Object value) {
+        if (data == null) {
+            data = new HashMap<>();
+        }
+
+        data.put(key, value);
+    }
+
+    @SuppressWarnings("unchecked")
+    public <T> Optional<T> get(String key, Class<T> klazz) {
+        return Optional.ofNullable((T) data.get(key));
+    }
+}