+++ src/test/java/ru/hh/consul/util/TrustManagerUtilsTest.java
@@ -1,0 +1,11 @@
+package ru.hh.consul.util;
+
+import static org.junit.Assert.assertNotNull;
+import org.junit.Test;
+
+public class TrustManagerUtilsTest {
+    @Test
+    public void shouldTrustManagerReturnCorrectResult() {
+        assertNotNull(TrustManagerUtils.getDefaultTrustManager());
+    }
+}